Did exercise 3

+++ submit/lab7/files/multiple-defs/f.c
@@ -1,4 +1,6 @@
-double x;
+#include "f.h"
+
+int x;
 
 void f() {
   x = -0.0;
